Add resource delete for serviceaccount, cluster-role, binding Split gke.go into 2 files for better readibility
